Hopefully working: Group Menu, Tracking
-Had some issues with GroupScene. Please utilise the scene GS.
This scene is missing some work that had been done for this testing session, but I should have them redone by the next test session.

-A group menu. Haven't done a directional arrow yet, I do have space for it and it would use the same script. Going to look into this before next meeting as well as some way to show that members are updated dynamically.

-To utilise this scene:
Interact with the group button on the hand menu that is not currently attached to the hand.
Select a group member to track them. (note, distance is not currently updated as you approach.)
Arrow was originally attached to the camera, however I have detached them as it's likely not how the arrow will display, I intend for it to be attached to the hand or something similar, depending on how the arrow is displayed for navigation.